Atualizando a funcao da pasta service
